refactor: 유효성 검사 부분을 도메인으로 이동
